Consolidated playblast utilities under `pipe/playblast/`
