+++ pipeline/pipe/h/playblast/__init__.py
@@ -1,5 +1,5 @@
-from .playblaster import HPlayblaster as HPlayblaster
-from .tool import launch_playblast as launch_playblast
+from pipe.h.playblast.launcher import launch_playblast as launch_playblast
+from pipe.h.playblast.playblaster import HPlayblaster as HPlayblaster
 
 __all__ = [
     "HPlayblaster",
