--- pipeline/pipe/h/playblast/constants.py
@@ -1,3 +1,0 @@
-from __future__ import annotations
-
-DEFAULT_RESOLUTION = (1280, 720)
