+++ pipeline/pipe/m/playblast/__init__.py
@@ -1,3 +1,13 @@
-from .anim import AnimPlayblastDialog as AnimPlayblastDialog
-from .previs import PrevisPlayblastDialog as PrevisPlayblastDialog
-from .turnaround import AssetTurnaroundDialog as AssetTurnaroundDialog
+from pipe.m.playblast.shot.anim import AnimPlayblastDialog as AnimPlayblastDialog
+from pipe.m.playblast.shot.previs import (
+    PrevisPlayblastDialog as PrevisPlayblastDialog,
+)
+from pipe.m.playblast.turnaround.dialog import (
+    AssetTurnaroundDialog as AssetTurnaroundDialog,
+)
+
+__all__ = [
+    "AnimPlayblastDialog",
+    "AssetTurnaroundDialog",
+    "PrevisPlayblastDialog",
+]
