+++ pipeline/pipe/m/playblast/shot/__init__.py
@@ -1,0 +1,6 @@
+"""Shot-playblast subsystem (anim + previs).
+
+The base `MPlayblastDialog` lives in `dialog.py` and is shared by
+`anim.AnimPlayblastDialog` and `previs.PrevisPlayblastDialog`. The
+`MPlayblaster` engine in `playblaster.py` drives the underlying capture.
+"""
