+++ pipeline/pipe/m/playblast/shot/playblaster.py
@@ -2,19 +2,18 @@ from __future__ import annotations
 
 import copy
 import logging
-from contextlib import contextmanager
 from typing import TYPE_CHECKING
 
 import maya.cmds as mc
 from mayacapture.capture import capture  # type: ignore[import-not-found]
 
+from pipe.m.playblast.hud import applied_hud
+from pipe.m.playblast.shot.config import MPlayblastConfig
 from pipe.m.util import maintain_selection
-from pipe.util import Playblaster
-
-from .struct import HudDefinition, MPlayblastConfig
+from pipe.playblast import Playblaster
 
 if TYPE_CHECKING:
-    from typing import Any, Generator
+    from typing import Any
 
 log = logging.getLogger(__name__)
 
@@ -137,51 +136,4 @@ class MPlayblaster(Playblaster):
                     )
 
 
-@contextmanager
-def applied_hud(
-    builtin_huds: list[str], custom_huds: list[HudDefinition]
-) -> Generator[None, None, None]:
-    # hide current huds and store current state
-    orig_visibility: dict[str, bool] = {}
-    orig_huds: list[str] = mc.headsUpDisplay(query=True, listHeadsUpDisplays=True)  # type: ignore
-    for hud in orig_huds:
-        vis = bool(mc.headsUpDisplay(hud, query=True, visible=True))
-        orig_visibility[hud] = vis
-        if vis:
-            mc.headsUpDisplay(hud, edit=True, visible=False)
-
-    # display requested builtin huds
-    for hud in builtin_huds:
-        mc.headsUpDisplay(hud, edit=True, visible=True)
-
-    # create requested custom huds
-    for chud in custom_huds:
-        if chud.name in orig_huds:
-            mc.headsUpDisplay(chud.name, remove=True)
-
-        kwargs: dict[str, Any] = dict()
-        if chud.idle_refresh:
-            kwargs.update({"attachToRefresh": True})
-        else:
-            kwargs.update({"event": chud.event})
-
-        mc.headsUpDisplay(
-            chud.name,
-            block=mc.headsUpDisplay(nextFreeBlock=chud.section),  # type: ignore
-            blockSize=chud.blockSize,
-            command=chud.command,
-            label=chud.label,
-            labelFontSize=chud.labelFontSize,
-            section=chud.section,
-            **kwargs,
-        )
-
-    try:
-        yield
-    finally:
-        # restore original visibility
-        for hud, state in orig_visibility.items():
-            mc.headsUpDisplay(hud, edit=True, visible=state)
-
-        for chud in custom_huds:
-            mc.headsUpDisplay(chud.name, remove=True)
+__all__ = ["MPlayblaster"]
