+++ pipeline/pipe/playblast/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from pipe.playblast.playblaster import Playblaster
+from pipe.playblast.presets import FFmpegPreset
+
+__all__ = [
+    "FFmpegPreset",
+    "Playblaster",
+]
